Reorganized classes, deleted excess code, and bagan modularizing for legibility and cohesion.
