+++ lib/screens/floating_buttons.dart
@@ -1,0 +1,40 @@
+// floating_buttons.dart
+import 'package:flutter/material.dart';
+
+/// Buttons the user can press to edit, view, or add to the Pie.
+class FloatingButtons extends StatelessWidget {
+  final bool editModeOn;
+  final VoidCallback toggleEditMode;
+  final VoidCallback showAddSliceDialog;
+
+  const FloatingButtons({
+    required this.editModeOn,
+    required this.toggleEditMode,
+    required this.showAddSliceDialog,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.end,
+      children: <Widget>[
+        Text(
+          editModeOn ? "Edit Mode is ON " : "Edit Mode is OFF ",
+          style: const TextStyle(fontSize: 24),
+        ),
+        FloatingActionButton(
+          onPressed: toggleEditMode,
+          tooltip: 'Toggle Edit Mode',
+          child: const Icon(Icons.edit),
+        ),
+        const SizedBox(width: 10),
+        FloatingActionButton(
+          onPressed: showAddSliceDialog,
+          tooltip: 'Add Slice',
+          child: const Icon(Icons.add),
+        ),
+      ],
+    );
+  }
+}
