+++ lib/screens/pie_provider.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+/// Manage the Pie's state.
+
+class PieProvider with ChangeNotifier {
+  bool _editModeOn = false;
+  bool get editModeOn => _editModeOn;
+
+  void toggleEditMode() {
+    _editModeOn = !_editModeOn;
+    notifyListeners();
+  }
+}
